web/requirements.txt
streamlit==1.52.2
requests==2.32.5
pydantic==2.12.5
python-magic==0.4.27
